exchange based algorithm
This is the algorithm approach that is based on exchange two abnormal
numbers.
- If we sort sausages in the order of length (or width), then we now
know the list is ordered by length (or width), but there are abnormal
orders by width (or length)
- So, the strategy is to replace two abnormal numbers into one by
exchanging, in fact replacing.
- We can exchanging numbers when below conditions are met.
-- even if remove the number (lower), the order shall be in order. That
is, if we want to move the number index i, then the width of index (i-1)
and index (i+1) shall be in order.
-- the place where the number will be put also be in order by the result
of exchange

Code description
f() : the function f handles the list of abnormal numbers. from the
first abnormal numbers it checks if it can be removed, and if yes, it
will search for the place it fits. once it's found, it will be put.
then, it will start from the next abnormal numbers again.

This function f() will be called until there will be no changes in the
number of abnormal numbers.

Result : this algorithm could not produce the right answer. it produced
far larger value than optimal value.